Movies fetch and rerender overwriting the loading div
I was trying to call a movie function and have it return an array
of li to render, and React wasn't having it. Changed the function to
Capitalized and then used a custom element which passed the movie
data in on the props object.

Unclear if the other way might have worked, but this renders.